Ensure Model ID is unique
The ID of the newly created model was obtained as the length of the
array of models, but if some models "in the middle" were deleted, the
new ID could be the same as the existing one.

We are sure that the last ID is the maximum, so we just take +1 than the
last element.